requirement.txt
admin-reports==0.10.8
aniso8601==7.0.0
annotated-types==0.5.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.2
asttokens==2.2.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.1.0
bootstrap-admin==0.4.4
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
django-admin-black==1.0.0
django-admin-extra-buttons==1.5.6
django-admin-report==0.6.2
django-admin-soft-dashboard==1.0.9
django-admin-views==1.0.3
django-appconf==1.0.5
django-auto-logout==0.5.0
django-axes==6.2.0
django-baton==2.6.0
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-crispy-forms==2.0
django-fontawesome==1.0
django-import-export==3.0.2
django-jazzmin==2.6.0
django-jquery==3.1.0
django-mathfilters==1.0.0
django-report-builder==6.4.2
django-reporting==0.21
django-reports-admin==1.1.0
django-select2==8.0.0
django-slick-reporting==0.6.7
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
docx2txt==0.8
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.19.1
fqdn==1.5.1
greenlet==2.0.1
idna==3.4
install==1.3.5
ipykernel==6.28.0
ipython==8.11.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.3
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.7.4
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.0
overrides==7.4.0
package_name==0.1
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
piexif==1.1.3
Pillow==9.3.0
placeholders==1.1
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.38
psutil==5.9.7
psycopg2-binary==2.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.12
PyYAML==6.0
pyzmq==25.1.2
referencing==0.32.1
regex==2022.10.31
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
Rx==1.6.3
scikit-learn==1.3.2
scipy==1.11.4
search-placeholder==0.2.0
Send2Trash==1.8.2
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.45
sqlparse==0.4.3
stack-data==0.6.2
tablib==3.3.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.64.1
traitlets==5.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.7.1
tzdata==2022.6
uri-template==1.3.0
urllib3==1.26.13
voila==0.5.5
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
xlrd==2.0.1
xlwt==1.3.0
